Add types to filter by category (and thus any field) for each entry type
best practice may be to create explicit types for each of the kinds of metadata (parts of speech, etymology, tone, context, topic, etc). but for now this is flexible and i’ll be one providing filters on front end so good enough and less code to write for now.

also worth noting, i got hung up on why resolver wasn’t working. i had to return array of Word objects [Word] in the Query type definition